scripts/requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Image Processing and GUI
Pillow>=9.5.0
tkinter-tooltip>=2.0.0

# Data Analysis and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Configuration and Data Handling
PyYAML>=6.0.0
pathlib2>=2.3.7

# Additional Dependencies for Advanced Features
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0

# Optional GPU Support (uncomment if using CUDA)
# torch-audio>=2.0.0
# torchtext>=0.15.0

# Development and Testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
